+++ common/rcontext/request_context.go
@@ -2,6 +2,7 @@ package rcontext
 
 import (
 	"context"
+	"net/http"
 
 	"github.com/sirupsen/logrus"
 	"github.com/turt2live/matrix-media-repo/common/config"
@@ -17,6 +18,7 @@ func Initial() RequestContext {
 			Thumbnails:        config.Get().Thumbnails.ThumbnailsConfig,
 			UrlPreviews:       config.Get().UrlPreviews.UrlPreviewsConfig,
 		},
+		Request: nil,
 	}.populate()
 }
 
@@ -24,13 +26,15 @@ type RequestContext struct {
 	context.Context
 
 	// These are also stored on the context object itself
-	Log    *logrus.Entry           // mr.logger
-	Config config.DomainRepoConfig // mr.serverConfig
+	Log     *logrus.Entry           // mr.logger
+	Config  config.DomainRepoConfig // mr.serverConfig
+	Request *http.Request           // mr.request
 }
 
 func (c RequestContext) populate() RequestContext {
 	c.Context = context.WithValue(c.Context, "mr.logger", c.Log)
-	//c.Context = context.WithValue(c.Context, "mr.serverConfig", c.Config)
+	c.Context = context.WithValue(c.Context, "mr.serverConfig", c.Config)
+	c.Context = context.WithValue(c.Context, "mr.request", c.Request)
 	return c
 }
 
@@ -40,6 +44,7 @@ func (c RequestContext) ReplaceLogger(log *logrus.Entry) RequestContext {
 		Context: ctx,
 		Log:     log,
 		Config:  c.Config,
+		Request: c.Request,
 	}
 }